backend(data): introduce data layer refactoring, STEP 1
